Refactor profile granularity code.
